third_party/blink/web_tests/external/wpt/css/css-grid/parsing/grid-area-valid-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['grid-column-start'] = "span calc(-2)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['grid-column-start'] = "span min(-1, 6)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['grid-column-start'] = "span calc(sibling-index() - 2)" should set the property value
  assert_equals: serialization should be canonical expected "span calc(sibling-index() - 2)" but got "span calc(-2 + sibling-index())"
Harness: the test ran to completion.
